requirements.txt
apache-airflow==2.8.1
requests
pandas
python-dotenv
sqlalchemy